Fix mingw build (bibref patch)

+++ src/modules/filters/cipherfil.cpp
@@ -56,7 +56,7 @@ char CipherFilter::processText(SWBuf &text, const SWKey *key, const SWModule *mo
 			text.setSize(len + 5);
 			memcpy(text.getRawData(), cipher->getUncipheredBuf(), len);
 		}
-		else if ((unsigned long)key == 1) {
+		else if ((size_t)key == 1) {
 			cipher->setUncipheredBuf(text.getRawData(), len);
 			cipher->getCipheredBuf(&len);
 			text.setSize(len + 5);
