+++ src/modules/filters/scsuutf8.cpp
@@ -139,7 +139,7 @@ int SCSUUTF8::UTF8Output(unsigned long uchar, SWBuf* utf8Buf)
 #endif
 
 char SCSUUTF8::processText(SWBuf &text, const SWKey *key, const SWModule *module) {
-	if ((unsigned long)key < 2)	// hack, we're en(1)/de(0)ciphering
+	if ((size_t)key < 2)	// hack, we're en(1)/de(0)ciphering
 		return -1;
 	
 #ifdef _ICU_
