+++ src/modules/filters/utf8nfc.cpp
@@ -42,7 +42,7 @@ UTF8NFC::~UTF8NFC() {
 
 char UTF8NFC::processText(SWBuf &text, const SWKey *key, const SWModule *module)
 {
-	if ((unsigned long)key < 2)	// hack, we're en(1)/de(0)ciphering
+	if ((size_t)key < 2)	// hack, we're en(1)/de(0)ciphering
 		return -1;
         
 	err = U_ZERO_ERROR;
